SLING-7970 Add Feature Model introspection service
Get the launcher to install an inventory printer service that prints the
effective feature. This can be used to report on the selected feature.
Additionally register a service that can find the bundle artifact ID
from a given Bundle Symbolic Name and Version. This can be used to find
a bundle in a feature description.