SLING-8334 Get Launcher Extension API ready for 1.0 release
The extensions now get provided with a single ExtensionContext object
that combines the functionality that was previously spread over the
LauncherPrepareContext and ExtensionInstallationContext.
Additionally, a lookup is provided for a Feature Model based on Artifact
ID. This can be used by extension plugins to read feature models
referenced in an extension.